+++ azote/common.py
@@ -54,3 +54,5 @@ settings = None         # object saved to / restored from ~/.azote/settings.pkl
 
 modes_swaybg = ["stretch", "fit", "fill", "center", "tile"]
 modes_feh = ["scale", "max", "fill", "center", "tile"]
+
+main_window = None